backend/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
Flask-Login==0.6.3
streamlit==1.32.2
requests==2.31.0
pandas==2.2.1
plotly==5.19.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5 